items: move Item_Animate to TRX
This makes the Item_Animate function common in TRX.
