+++ src/libtrx/game/items.c
@@ -1,6 +1,7 @@
 #include "game/items.h"
 
 #include "game/const.h"
+#include "game/item_actions.h"
 #include "game/lara/common.h"
 #include "game/objects/common.h"
 #include "game/objects/vars.h"
@@ -121,6 +122,107 @@ void Item_Translate(
     item->pos.x += ((c * x + s * z) >> W2V_SHIFT);
     item->pos.y += y;
     item->pos.z += ((c * z - s * x) >> W2V_SHIFT);
+}
+
+void Item_Animate(ITEM *const item)
+{
+    item->hit_status = 0;
+    item->touch_bits = 0;
+    item->frame_num++;
+
+    const ANIM *anim = Item_GetAnim(item);
+    if (anim->num_changes > 0 && Item_GetAnimChange(item, anim)) {
+        anim = Item_GetAnim(item);
+        item->current_anim_state = anim->current_anim_state;
+
+        if (item->required_anim_state == anim->current_anim_state) {
+            item->required_anim_state = 0;
+        }
+    }
+
+    if (item->frame_num > anim->frame_end) {
+        for (int32_t i = 0; i < anim->num_commands; i++) {
+            const ANIM_COMMAND *const command = &anim->commands[i];
+            switch (command->type) {
+            case AC_MOVE_ORIGIN: {
+                const XYZ_16 *const pos = (XYZ_16 *)command->data;
+                Item_Translate(item, pos->x, pos->y, pos->z);
+                break;
+            }
+
+            case AC_JUMP_VELOCITY: {
+                const ANIM_COMMAND_VELOCITY_DATA *const data =
+                    (ANIM_COMMAND_VELOCITY_DATA *)command->data;
+                item->fall_speed = data->fall_speed;
+                item->speed = data->speed;
+                item->gravity = true;
+                break;
+            }
+
+            case AC_DEACTIVATE:
+                item->status = IS_DEACTIVATED;
+                break;
+
+            default:
+                break;
+            }
+        }
+
+        item->anim_num = anim->jump_anim_num;
+        item->frame_num = anim->jump_frame_num;
+        anim = Item_GetAnim(item);
+
+#if TR_VERSION == 1
+        item->current_anim_state = anim->current_anim_state;
+        item->goal_anim_state = item->current_anim_state;
+#else
+        if (item->current_anim_state != anim->current_anim_state) {
+            item->current_anim_state = anim->current_anim_state;
+            item->goal_anim_state = anim->current_anim_state;
+        }
+#endif
+
+        if (item->required_anim_state == item->current_anim_state) {
+            item->required_anim_state = 0;
+        }
+    }
+
+    for (int32_t i = 0; i < anim->num_commands; i++) {
+        const ANIM_COMMAND *const command = &anim->commands[i];
+        switch (command->type) {
+        case AC_SOUND_FX: {
+            const ANIM_COMMAND_EFFECT_DATA *const data =
+                (ANIM_COMMAND_EFFECT_DATA *)command->data;
+            Item_PlayAnimSFX(item, data);
+            break;
+        }
+
+        case AC_EFFECT:
+            const ANIM_COMMAND_EFFECT_DATA *const data =
+                (ANIM_COMMAND_EFFECT_DATA *)command->data;
+            if (item->frame_num == data->frame_num) {
+                ItemAction_Run(data->effect_num, item);
+            }
+            break;
+
+        default:
+            break;
+        }
+    }
+
+    if (item->gravity != 0) {
+        item->fall_speed += item->fall_speed < FAST_FALL_SPEED ? GRAVITY : 1;
+        item->pos.y += item->fall_speed;
+    } else {
+        int32_t speed = anim->velocity;
+        if (anim->acceleration != 0) {
+            speed += anim->acceleration * (item->frame_num - anim->frame_base);
+        }
+        item->speed = speed >> 16;
+    }
+
+    item->pos.x += (item->speed * Math_Sin(item->rot.y)) >> W2V_SHIFT;
+    item->pos.z += (item->speed * Math_Cos(item->rot.y)) >> W2V_SHIFT;
 }
 
 void Item_PlayAnimSFX(
