+++ src/tr1/game/items.c
@@ -3,7 +3,6 @@
 #include "game/carrier.h"
 #include "game/effects.h"
 #include "game/interpolation.h"
-#include "game/item_actions.h"
 #include "game/random.h"
 #include "game/room.h"
 #include "game/shell.h"
@@ -543,104 +542,6 @@ void Item_ShiftCol(ITEM *item, COLL_INFO *coll)
     coll->shift.x = 0;
     coll->shift.y = 0;
     coll->shift.z = 0;
-}
-
-void Item_Animate(ITEM *item)
-{
-    item->touch_bits = 0;
-    item->hit_status = 0;
-
-    const ANIM *anim = Item_GetAnim(item);
-
-    item->frame_num++;
-
-    if (anim->num_changes > 0) {
-        if (Item_GetAnimChange(item, anim)) {
-            anim = Item_GetAnim(item);
-            item->current_anim_state = anim->current_anim_state;
-
-            if (item->required_anim_state == item->current_anim_state) {
-                item->required_anim_state = 0;
-            }
-        }
-    }
-
-    if (item->frame_num > anim->frame_end) {
-        for (int32_t i = 0; i < anim->num_commands; i++) {
-            const ANIM_COMMAND *const command = &anim->commands[i];
-            switch (command->type) {
-            case AC_MOVE_ORIGIN: {
-                const XYZ_16 *const pos = (XYZ_16 *)command->data;
-                Item_Translate(item, pos->x, pos->y, pos->z);
-                break;
-            }
-
-            case AC_JUMP_VELOCITY: {
-                const ANIM_COMMAND_VELOCITY_DATA *const data =
-                    (ANIM_COMMAND_VELOCITY_DATA *)command->data;
-                item->fall_speed = data->fall_speed;
-                item->speed = data->speed;
-                item->gravity = true;
-                break;
-            }
-
-            case AC_DEACTIVATE:
-                item->status = IS_DEACTIVATED;
-                break;
-
-            default:
-                break;
-            }
-        }
-
-        item->anim_num = anim->jump_anim_num;
-        item->frame_num = anim->jump_frame_num;
-
-        anim = Item_GetAnim(item);
-        item->current_anim_state = anim->current_anim_state;
-        item->goal_anim_state = item->current_anim_state;
-        if (item->required_anim_state == item->current_anim_state) {
-            item->required_anim_state = 0;
-        }
-    }
-
-    for (int32_t i = 0; i < anim->num_commands; i++) {
-        const ANIM_COMMAND *const command = &anim->commands[i];
-        switch (command->type) {
-        case AC_SOUND_FX: {
-            const ANIM_COMMAND_EFFECT_DATA *const data =
-                (ANIM_COMMAND_EFFECT_DATA *)command->data;
-            Item_PlayAnimSFX(item, data);
-            break;
-        }
-
-        case AC_EFFECT: {
-            const ANIM_COMMAND_EFFECT_DATA *const data =
-                (ANIM_COMMAND_EFFECT_DATA *)command->data;
-            if (item->frame_num == data->frame_num) {
-                ItemAction_Run(data->effect_num, item);
-            }
-            break;
-        }
-
-        default:
-            break;
-        }
-    }
-
-    if (!item->gravity) {
-        int32_t speed = anim->velocity;
-        if (anim->acceleration) {
-            speed += anim->acceleration * (item->frame_num - anim->frame_base);
-        }
-        item->speed = speed >> 16;
-    } else {
-        item->fall_speed += (item->fall_speed < FAST_FALL_SPEED) ? GRAVITY : 1;
-        item->pos.y += item->fall_speed;
-    }
-
-    item->pos.x += (Math_Sin(item->rot.y) * item->speed) >> W2V_SHIFT;
-    item->pos.z += (Math_Cos(item->rot.y) * item->speed) >> W2V_SHIFT;
 }
 
 bool Item_IsTriggerActive(ITEM *item)
