+++ src/tr2/game/items.h
@@ -23,7 +23,6 @@ int32_t Item_TestPosition(
     const int16_t *bounds, const ITEM *src_item, const ITEM *dst_item);
 void Item_AlignPosition(
     const XYZ_32 *vec, const ITEM *src_item, ITEM *dst_item);
-void Item_Animate(ITEM *item);
 int32_t Item_IsTriggerActive(ITEM *item);
 int32_t Item_GetFrames(const ITEM *item, ANIM_FRAME *frmptr[], int32_t *rate);
 BOUNDS_16 *Item_GetBoundsAccurate(const ITEM *item);
